refactor: Dashboard page + own components
