+++ src/components/Dashboard/ApisTable/index.ts
@@ -1,0 +1,1 @@
+export {default as ApisTable} from './ApisTable';
