+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export {AddServerModal} from './AddServerModal';
-export {ApiInfo} from './ApiInfo';
-export {Dashboard} from './Dashboard';
+export {ApiInfo, ApisTable} from './Dashboard';
+export {EnvoyFleetInfoModal} from './EnvoyFleetInfoModal';
 export {PostProcessedApiSpec} from './PostProcessedApiSpec';
 export {RawApiSpec} from './RawApiSpec';
 export {Sidebar} from './Sidebar';
