+++ src/models/dashboard.ts
@@ -1,4 +1,4 @@
-export interface DashboardAPIsTableDataSourceItem {
+export interface ApisTableDataSourceItem {
   key: string;
   name: string;
   services: 'available' | 'unavailable';
