+++ src/pages/Dashboard/index.ts
@@ -1,0 +1,1 @@
+export {default as Dashboard} from './Dashboard';
