+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export {Dashboard} from './Dashboard';
+export {Settings} from './Settings';
